requirements.txt
beautifulsoup4==4.11.2
Django==4.1.5
django_ckeditor==6.5.1
django_cloudinary_storage==0.3.0
django_tinymce==3.5.0
django-material-admin==1.8.6
